Move Node command into its own file
